Restrict OAuth SSO login to University of Cambridge accounts
This commit updates the OAuth SSO integration in our Pretix instance to
restrict logins to University of Cambridge accounts. It hard-codes the
addition of the `hd` parameter in the OAuth request to ensure only
accounts with the University of Cambridge domain (e.g., @cam.ac.uk) can
authenticate. Additionally, the response is checked to verify the
presence of this claim.

This implementation is specific to our configuration and will cause
issues if other OAuth providers are used. A more generic solution
cannot be implemented until pluggable SSO backends are supported, which
has not yet been introduced:

  https://github.com/pretix/pretix/discussions/3469