Display full fee description during checkout
By default, Pretix only displays the fee type to users during checkout,
omitting the fee description. This behaviour appears to be intentional,
as discussed here:

  https://github.com/pretix/pretix/discussions/3782#discussioncomment-8001444

This commit modifies this behaviour to show both the fee type and
description, if a description is available, providing users with more
complete information during the checkout process.